cleanup sort changes
