+++ src/channel.ts
@@ -1,7 +1,7 @@
 import Immutable from 'seamless-immutable';
 import { ChannelState } from './channel_state';
 import { isValidEventType } from './events';
-import { logChatPromiseExecution } from './utils';
+import { logChatPromiseExecution, normalizeQuerySort } from './utils';
 import { StreamChat } from './client';
 import {
   APIResponse,
@@ -355,7 +355,7 @@ export class Channel<
    * queryMembers - Query Members
    *
    * @param {UserFilters<UserType>}  filterConditions object MongoDB style filters
-   * @param {UserSort<UserType> | UserSort<UserType>[]} [sort] Sort options, for instance {created_at: -1}.
+   * @param {UserSort<UserType>} [sort] Sort options, for instance {created_at: -1}.
    * When using multiple fields, make sure you use array of objects to guarantee field order, for instance [{last_active: -1}, {created_at: 1}]
    * @param {{ limit?: number; offset?: number }} [options] Option object, {limit: 10, offset:10}
    *
@@ -363,7 +363,7 @@ export class Channel<
    */
   async queryMembers(
     filterConditions: UserFilters<UserType>,
-    sort: UserSort<UserType> | UserSort<UserType>[] = [],
+    sort: UserSort<UserType> = [],
     options: { limit?: number; offset?: number } = {},
   ) {
     let id: string | undefined;
@@ -382,7 +382,7 @@ export class Channel<
           type,
           id,
           members,
-          sort: this.getClient()._normalizeQuerySort(sort),
+          sort: normalizeQuerySort(sort),
           filter_conditions: filterConditions,
           ...options,
         },
